11. ORM class들을 패키지에 따로 모아두기
